+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v1.2.2 - (Unreleased)
 ----
  * Added support to use collection filters for interfaces (polymorphic nodes)
  * Added support to set filters type with modifiers `[]` or `!` 
+ * Added support to use page based pagination
   
 v1.2.1 - 2018-07-16
 ----